Implementing necessary DocumentParser classes.
Because this application will be used with various countries and banks.
Each of country's banks are having their own standard of formatation of documents, some may even not using CSV file.
Therefore, the Parser handler must be aware of this incompatible of format.
Meaning that we should decoupling the responsibility of parser and file formatting as much as possible.
That would allow the code to be flexible enough to add or remove new format, new file type.